Put bitshuffle functions in a new class file
