+++ src/main/java/org/xerial/snappy/Snappy.java
@@ -903,174 +903,4 @@ public class Snappy
         byte[] uncompressed = uncompress(input);
         return new String(uncompressed, encoding);
     }
-
-    /**
-     * Apply a bit-shuffling filter into the input short array.
-     *
-     * @param input
-     * @return bit-shuffled byte array
-     * @throws IOException
-     */
-    public static byte[] bitShuffle(short[] input) throws IOException {
-        byte[] output = new byte[input.length * 2];
-        if (impl.supportBitSuffle()) {
-            impl.bitShuffle(input, 0, 2, input.length * 2, output, 0);
-        } else {
-            arrayCopy(input, 0, input.length * 2, output, 0);
-        }
-        return output;
-    }
-
-    /**
-     * Apply a bit-shuffling filter into the input int array.
-     *
-     * @param input
-     * @return bit-shuffled byte array
-     * @throws IOException
-     */
-    public static byte[] bitShuffle(int[] input) throws IOException {
-        byte[] output = new byte[input.length * 4];
-        if (impl.supportBitSuffle()) {
-            impl.bitShuffle(input, 0, 4, input.length * 4, output, 0);
-        } else {
-            arrayCopy(input, 0, input.length * 4, output, 0);
-        }
-        return output;
-    }
-
-    /**
-     * Apply a bit-shuffling filter into the input long array.
-     *
-     * @param input
-     * @return bit-shuffled byte array
-     * @throws IOException
-     */
-    public static byte[] bitShuffle(long[] input) throws IOException {
-        byte[] output = new byte[input.length * 8];
-        if (impl.supportBitSuffle()) {
-            impl.bitShuffle(input, 0, 8, input.length * 8, output, 0);
-        } else {
-            arrayCopy(input, 0, input.length * 8, output, 0);
-        }
-        return output;
-    }
-
-    /**
-     * Apply a bit-shuffling filter into the input float array.
-     *
-     * @param input
-     * @return bit-shuffled byte array
-     * @throws IOException
-     */
-    public static byte[] bitShuffle(float[] input) throws IOException {
-        byte[] output = new byte[input.length * 4];
-        if (impl.supportBitSuffle()) {
-            impl.bitShuffle(input, 0, 4, input.length * 4, output, 0);
-        } else {
-            arrayCopy(input, 0, input.length * 4, output, 0);
-        }
-        return output;
-    }
-
-    /**
-     * Apply a bit-shuffling filter into the input double array.
-     *
-     * @param input
-     * @return bit-shuffled byte array
-     * @throws IOException
-     */
-    public static byte[] bitShuffle(double[] input) throws IOException {
-        byte[] output = new byte[input.length * 8];
-        if (impl.supportBitSuffle()) {
-            impl.bitShuffle(input, 0, 8, input.length * 8, output, 0);
-        } else {
-            arrayCopy(input, 0, input.length * 8, output, 0);
-        }
-        return output;
-    }
-
-    /**
-     * Convert the input bit-shuffled byte array into an original short array.
-     *
-     * @param input
-     * @return a short array
-     * @throws IOException
-     */
-    public static short[] bitUnShuffleShortArray(byte[] input) throws IOException {
-        short[] output = new short[input.length / 2];
-        if (impl.supportBitSuffle()) {
-            impl.bitUnShuffle(input, 0, 2, input.length, output, 0);
-        } else {
-            arrayCopy(input, 0, input.length, output, 0);
-        }
-        return output;
-    }
-
-    /**
-     * Convert the input bit-shuffled byte array into an original int array.
-     *
-     * @param input
-     * @return an int array
-     * @throws IOException
-     */
-    public static int[] bitUnShuffleIntArray(byte[] input) throws IOException {
-        int[] output = new int[input.length / 4];
-        if (impl.supportBitSuffle()) {
-            impl.bitUnShuffle(input, 0, 4, input.length, output, 0);
-        } else {
-            arrayCopy(input, 0, input.length, output, 0);
-        }
-        return output;
-    }
-
-    /**
-     * Convert the input bit-shuffled byte array into an original long array.
-     *
-     * @param input
-     * @return a long array
-     * @throws IOException
-     */
-    public static long[] bitUnShuffleLongArray(byte[] input) throws IOException {
-        long[] output = new long[input.length / 8];
-        if (impl.supportBitSuffle()) {
-            impl.bitUnShuffle(input, 0, 8, input.length, output, 0);
-        } else {
-            arrayCopy(input, 0, input.length, output, 0);
-        }
-        return output;
-    }
-
-    /**
-     * Convert the input bit-shuffled byte array into an original float array.
-     *
-     * @param input
-     * @return an float array
-     * @throws IOException
-     */
-    public static float[] bitUnShuffleFloatArray(byte[] input) throws IOException {
-        float[] output = new float[input.length / 4];
-        if (impl.supportBitSuffle()) {
-            impl.bitUnShuffle(input, 0, 4, input.length, output, 0);
-        } else {
-            arrayCopy(input, 0, input.length, output, 0);
-        }
-        return output;
-    }
-
-    /**
-     * Convert the input bit-shuffled byte array into an original double array.
-     *
-     * @param input
-     * @return a double array
-     * @throws IOException
-     */
-    public static double[] bitUnShuffleDoubleArray(byte[] input) throws IOException {
-        double[] output = new double[input.length / 8];
-        if (impl.supportBitSuffle()) {
-            impl.bitUnShuffle(input, 0, 8, input.length, output, 0);
-        } else {
-            arrayCopy(input, 0, input.length, output, 0);
-        }
-        return output;
-    }
 }
